almost got the new UserCommands + rc.__config__ playing nicely. Just missing OpticsManager
