+++ scopesim/commands/__init__.py
@@ -1,2 +1,1 @@
 from .user_commands import UserCommands
-from .user_commands_utils import *
