+++ scopesim/server/__init__.py
@@ -1,1 +1,1 @@
-from .OLD_database import *
+
