+++ scopesim/tests/mocks/files/CMD_unity_cmds.yaml
@@ -1,0 +1,24 @@
+alias : SIM
+properties :
+  spectral:
+    lam_min : 0.5     # um
+    lam_mid : 1.5     # um
+    lam_max : 2.5     # um
+
+  sub_pixel :
+    flag : True       # Turn on subpixel shifts for point sources
+    fraction : 1      # full pixel shift
+
+  computing :
+    chunk_size : 32  # pixel side length of per FOV
+
+---
+
+alias : OBS
+yamls : ["unity_sys.yaml"]
+
+---
+
+alias : INST
+properties :
+  pixel_scale : 1   # arcsec / pixel
