+++ test_package/default.yaml
@@ -1,0 +1,15 @@
+# Instrument
+object : observation
+alias : OBS
+name : test_instrument
+
+packages :
+- test_package
+
+yamls :
+- test_telescope.yaml
+- test_instrument.yaml
+- test_detector.yaml
+
+properties :
+  airmass : 1.
